Consolidate TrackPopupMenuEx in User32
