--- src/ManagedShell.ShellFolders/Enums/TPM.cs
@@ -1,29 +1,0 @@
-﻿using System;
-
-namespace ManagedShell.ShellFolders.Enums
-{
-    // Specifies how TrackPopupMenuEx positions the shortcut menu horizontally
-    [Flags]
-    public enum TPM : uint
-    {
-        LEFTBUTTON = 0x0000,
-        RIGHTBUTTON = 0x0002,
-        LEFTALIGN = 0x0000,
-        CENTERALIGN = 0x0004,
-        RIGHTALIGN = 0x0008,
-        TOPALIGN = 0x0000,
-        VCENTERALIGN = 0x0010,
-        BOTTOMALIGN = 0x0020,
-        HORIZONTAL = 0x0000,
-        VERTICAL = 0x0040,
-        NONOTIFY = 0x0080,
-        RETURNCMD = 0x0100,
-        RECURSE = 0x0001,
-        HORPOSANIMATION = 0x0400,
-        HORNEGANIMATION = 0x0800,
-        VERPOSANIMATION = 0x1000,
-        VERNEGANIMATION = 0x2000,
-        NOANIMATION = 0x4000,
-        LAYOUTRTL = 0x8000
-    }
-}
